+++ core/src/backgammon.py
@@ -4,6 +4,15 @@ from core.src.jugador import Jugador
 from abc import ABC, abstractmethod
 from enum import Enum, auto
 from typing import List # Para el type hinting
+import random # Para la tirada inicial
+from core.src.exceptions import (
+    PosicionOcupadaException, 
+    PrimerCuadranteIncompletoException, 
+    PosicionVaciaException,
+    MovimientoNoPosibleException,
+    TiradaInicialEmpateException,
+    NoEsMomentoException
+)
 
 class EstadoJuego(Enum):
     """Define los estados posibles del juego para la máquina de estados."""
@@ -65,6 +74,49 @@ class BackgammonGame:
     def ganador(self) -> Jugador:
         """Devuelve el ganador si es que lo hay"""
         return self.__ganador__
+
+    def lanzar_dados(self) -> None:
+        """Tira los dados, si es el primer turno, decide quien empieza, si no, es una tirada normal.
+        
+        Raises
+        ------
+        NoEsMomentoException
+            El estado del juego no es tirar los dados
+        TiradaInicialEmpateException
+            Se dio un empate en la tirada inicial
+        """
+
+        if self.__estado_juego__ != EstadoJuego.TIRANDO_DADOS:
+            raise NoEsMomentoException('No es el momento de tirar los dados')
+
+        if self.__primer_turno__ == True:
+            dado_j1 = random.randint(1, 6)
+            dado_j2 = random.randint(1, 6)
+
+            if dado_j1 == dado_j2:
+                raise TiradaInicialEmpateException('Empate, hay que tirar de nuevo')
+            elif dado_j1 > dado_j2:
+                self.__turno__ = self.__jugador1__
+            else:
+                self.__turno__ = self.__jugador2__
+
+            self.__dados__.set_valores([dado_j1, dado_j2])
+            self.__primer_turno__ = False
+            self.__estado_juego__ = EstadoJuego.MOVIENDO
+
+            return {
+                'dado_j1': dado_j1,
+                'dado_j2': dado_j2,
+                'ganador': self.__turno__.__nombre__
+            }
+        else:
+            self.__dados__.tirar()
+            self.__estado_juego__ = EstadoJuego.MOVIENDO
+            return {
+                'dados': self.__dados__.valores
+            }
+
+